Add first cut at an idea for a model.
Outstanding question: if you query for all KeyBase entities with a
reversed fingerprint with a certain value, and a PublicSubkey has that
value, what would be returned if the PublicSubkey were stored as a
StructuredProperty in a PublicKey entity?  What about if I used an
ancestor relationship instead?  In that case I know I could pop the child off
the key to fetch the containing PublicKey, but that's another roundtrip.
Better to potentially make 2 roundtrips, or to potentially have to OR
PublicKey.rfp and PublicKey.subkey.rfp index checks in the query?