Add code to generate ids used by WKD draft 5.
These are the result of extracting the localpart of the email address,
converting ascii letters to lowercase, computing the sha1 hash of that,
and then converting that digest to z-base-32 (a variant of base32 with a
different alphabet and no padding).

Note that my implementation of z-base-32 does not handle the case where
the input is not a multiple of 8 bits long.  This does not occur in the
usage in WKD, but if we ever need a general implementation of z-base-32,
a bit length param should be added and support for this should be
implemented.